Move healthcheck to middleware/pkg/healthcheck
